backend/requirements.txt
bcrypt==4.0.1
Flask==2.2.4
Flask-Cors==3.0.10
Flask-Security==3.0.0       
Flask-SQLAlchemy==3.0.3
Flask-RESTful==0.3.9
celery[redis]
email-validator==2.0.0.post2
flask_caching
Pillow==9.5.0